requirements.txt

google-adk

google-cloud-aiplatform[adk,agent_engines]>=1.38.0

google-cloud-bigquery>=3.11.0

google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

pandas>=1.5.0

python-dotenv>=1.0.0

colorlog>=6.7.0

absl-py>=2.2.1,<3.0.0
vertexai>=1.91.0
google-genai>=1.5.0,<2.0.0
pydantic>=2.10.6,<3.0.0